+++ app/scripts/backend/common.ts
@@ -1,5 +1,5 @@
 /** @format */
-import { httpConfAddMethodFetch } from "@/util"
+import { fetchConfAddMethod } from "@/util"
 import { getAuthorizationHeader } from "@/components/auth/auth"
 import settings from "@/settings"
 import { API, Response } from "./types"
@@ -8,7 +8,7 @@ export async function korpRequest<K extends keyof API>(
     endpoint: K,
     params: API[K]["params"]
 ): Promise<Response<API[K]["response"]>> {
-    const { url, request } = httpConfAddMethodFetch(settings.korp_backend_url + "/" + endpoint, params)
+    const { url, request } = fetchConfAddMethod(settings.korp_backend_url + "/" + endpoint, params)
     request.headers = { ...request.headers, ...getAuthorizationHeader() }
     const response = await fetch(url, request)
     return (await response.json()) as Response<API[K]["response"]>